# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是 Kano 系统的通用上传服务，支持多种云存储（本地、腾讯云、阿里云）。",
        "title": "Kano 上传服务 API",
        "termsOfService": "https://kano.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://kano.com/support",
            "email": "support@kano.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/kano/v1",
    "paths": {
        "/v1/upload/record": {
            "post": {
                "description": "将上传完成后的文件记录保存到数据库",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "上传服务"
                ],
                "summary": "保存上传记录",
                "parameters": [
                    {
                        "description": "上传记录数组",
                        "name": "records",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.UploadRecord"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "2001": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/v1/upload/token": {
            "get": {
                "description": "根据应用码获取上传凭证",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "上传服务"
                ],
                "summary": "获取上传凭证",
                "parameters": [
                    {
                        "type": "string",
                        "description": "应用码",
                        "name": "application_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {}
                                    }
                                }
                            ]
                        }
                    },
                    "2001": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "description": "使用interface{}以支持不同类型的数据"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.UploadRecord": {
            "type": "object",
            "properties": {
                "bucket_spec": {
                    "description": "bucket 规范",
                    "type": "string"
                },
                "business_type": {
                    "description": "用于区分不同模块的上传操作（如头像上传、素材上传等）",
                    "type": "integer"
                },
                "created_at": {
                    "description": "记录创建时间，默认当前时间",
                    "type": "string"
                },
                "credential_type": {
                    "description": "凭证类型（1普通凭证、2加密凭证）",
                    "type": "boolean"
                },
                "error_message": {
                    "description": "上传失败时的错误信息，用于定位问题原因，如“文件过大”、“格式不支持”、“网络中断”等",
                    "type": "string"
                },
                "file_path": {
                    "description": "文件在服务器上的存储路径，仅在上传成功时填写，用于后续访问文件或生成下载链接",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识每条上传记录，自增",
                    "type": "integer"
                },
                "status": {
                    "description": "0success, 1failure 表示该次上传是成功还是失败，用于判断是否需要触发告警或后续处理逻辑",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "upload_id": {
                    "description": "上传任务ID",
                    "type": "string"
                },
                "username": {
                    "description": "用户名字",
                    "type": "string"
                }
            }
        }
    }
}